+++ templates/user-lists/list-h2o.php
@@ -1,4 +1,0 @@
-<div
-	class="user-list">
-	<?php echo $users; ?>
-</div>